backend/requirements.txt
fastapi
uvicorn
mss
pillow
pytesseract
openai
python-dotenv
httpx
dashscope
pyautogui
pygetwindow
